Refactor pages folder
